requirements.txt
requests
pynput
PyQt6
python-dotenv
evdev
